% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scenario_prices.R
\name{apply_scenario_prices}
\alias{apply_scenario_prices}
\title{Calculate scenario prices}
\usage{
apply_scenario_prices(data, shock_year)
}
\arguments{
\item{shock_year}{Year of shock.}

\item{price_data}{A tibble holding price data.}
}
\value{
A tibble holding late_and_sudden_prices
}
\description{
Function generates prices for baseline and late and sudden shock scenario.
Price for baseline scenario correspond to prices of \code{baseline_scenario}.
Prices for the late sudden scenario also correspond to \code{baseline_scenario}
until the \code{year_of_shock}. From then on they linearly approach the price
level of the \code{shock_scenario} during the \code{duration_of_shock}.
}
