% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_profts.R
\name{calculate_annual_profits}
\alias{calculate_annual_profits}
\title{Calculate annual profits}
\usage{
calculate_annual_profits(
  data,
  baseline_scenario,
  shock_scenario,
  start_year,
  end_year,
  discount_rate,
  growth_rate
)
}
\arguments{
\item{data}{data frame containing the full trajectory company data}

\item{baseline_scenario}{Character. A string that indicates which
of the scenarios included in the analysis should be used to set the
baseline technology trajectories.}

\item{shock_scenario}{Character. A string that indicates which
of the scenarios included in the analysis should be used to set the
late & sudden technology trajectories.}

\item{start_year}{Numeric, holding start year of analysis. Used to pull the
terminal value back to present value as per the 2DII "Limited Visibility"
methodology (Figure 1).}

\item{end_year}{Numeric, holding end year of analysis.}

\item{discount_rate}{Numeric, the discount rate}

\item{growth_rate}{Numeric, that holds the terminal growth rate of profits
beyond the \code{end_year} in the DCF.}
}
\value{
A tibble holding annual profits
}
\description{
Wrapper function to calculate discounted annual profits and terminal value.
}
