% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{check_valid_financial_data_values}
\alias{check_valid_financial_data_values}
\title{Check if values in financial data are plausible}
\usage{
check_valid_financial_data_values(financial_data)
}
\arguments{
\item{financial_data}{data frame that is to be validated}
}
\description{
Checks that numeric columns hold values in acceptable ranges.
}
