% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trisk_trajectory.R
\name{extend_assets_trajectories}
\alias{extend_assets_trajectories}
\title{Calculate baseline and transition shock trajectoroes}
\usage{
extend_assets_trajectories(trisk_model_input, start_year, shock_year)
}
\arguments{
\item{trisk_model_input}{List with project agnostic and project specific input data}

\item{start_year}{Numeric, holding start year of analysis.}

\item{shock_year}{year of transition from baseline scenario to target scenario}
}
\value{
A tibble holding annual profits
}
\description{
Calculate baseline and transition shock trajectoroes
}
