% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{financial_features_testdata}
\alias{financial_features_testdata}
\title{Financial Features Test Data}
\format{
A data frame with the following columns:
\describe{
\item{company_id}{Unique identifier for the company.}
\item{pd}{Probability of default for the company.}
\item{net_profit_margin}{Net profit margin for the company.}
\item{debt_equity_ratio}{Debt to equity ratio.}
\item{volatility}{Volatility of the company's asset values.}
}
}
\usage{
financial_features_testdata
}
\description{
This dataset contains financial metrics necessary for calculating stress test outputs.
}
\note{
This dataset is accessible via \code{trisk.model:::financial_features_testdata} for testing purposes.
}
\keyword{datasets}
