% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{get_end_year}
\alias{get_end_year}
\title{Get End year from data}
\usage{
get_end_year(data, scenarios_filter, MAX_POSSIBLE_YEAR = 2050)
}
\arguments{
\item{data}{data}

\item{scenarios_filter}{scenarios to use}
}
\value{
the end year
}
\description{
Get End year from data
}
