% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{ngfs_carbon_price_testdata}
\alias{ngfs_carbon_price_testdata}
\title{NGFS Carbon Price Test Data}
\format{
A data frame with the following columns:
\describe{
\item{year}{Year of the carbon price.}
\item{model}{Model used to generate the carbon price.}
\item{scenario}{Scenario name.}
\item{scenario_geography}{Geographic region for the scenario.}
\item{variable}{The variable measured (e.g., carbon price).}
\item{unit}{Unit of the variable.}
\item{carbon_tax}{The amount of carbon tax applied in the scenario.}
}
}
\usage{
ngfs_carbon_price_testdata
}
\description{
This dataset provides carbon pricing data used in the stress test scenarios.
}
\note{
This dataset is accessible via \code{trisk.model:::carbon_price_testdata} for testing purposes.
}
\keyword{datasets}
