% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{prepare_trisk_results}
\alias{prepare_trisk_results}
\title{Prepare TRISK Results}
\usage{
prepare_trisk_results(output_list, trisk_params, run_id)
}
\arguments{
\item{output_list}{A list containing the output data from the TRISK model run.}

\item{trisk_params}{A data structure (e.g., list or tibble) containing the TRISK model parameters.}

\item{run_id}{A unique identifier for the model run.}
}
\value{
A list of tibbles containing the following elements:
\describe{
\item{npv}{A tibble with the NPV results of the TRISK model run.}
\item{pd}{A tibble with the PD results of the TRISK model run.}
\item{trajectories}{A tibble with the company trajectory results.}
\item{params}{A tibble with the model parameters used in the run.}
}
}
\description{
This function prepares a list of results from a TRISK model run, including net present value (NPV),
probability of default (PD), company trajectories, and model parameters. It converts these results into tibbles for further analysis.
}
