% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{process_params}
\alias{process_params}
\title{Process function parameters}
\usage{
process_params(fun, ...)
}
\arguments{
\item{fun}{A function whose parameters are to be processed.}

\item{...}{User-defined arguments to override the function's default parameters.}
}
\value{
A list of final parameters.
}
\description{
This function extracts the default parameters from a function and replaces them with user-provided values where applicable.
/!\ It ignore params without default, or with a NULL default
}
