% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{retype_data}
\alias{retype_data}
\title{Retype Data Columns}
\usage{
retype_data(assets_data, scenarios_data, financial_data, carbon_data)
}
\arguments{
\item{assets_data}{A dataframe containing assets data.}

\item{scenarios_data}{A dataframe containing scenarios data.}

\item{financial_data}{A dataframe containing financial data.}

\item{carbon_data}{A dataframe containing carbon price data.}
}
\value{
A list with retyped dataframes: \code{assets_data}, \code{scenarios_data},
\code{financial_data}, and \code{carbon_data}.
}
\description{
This function retypes the columns of assets_data, scenarios_data,
financial_data, and carbon_data according to their intended column types.
}
