% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{run_trisk}
\alias{run_trisk}
\title{Run stress testing for provided asset type}
\usage{
run_trisk(input_path, output_path = NULL, show_params_cols = TRUE, ...)
}
\arguments{
\item{input_path}{String containing the path to project-agnostic data.}

\item{output_path}{String containing the path to which output files are written.
NOTE: Results and logs per run are saved to a subdirectory of output_path
that will be generated automatically. The name of the subdirectory is the
timestamp of the run of the analysis.}
}
\value{
If \code{output_path} is NULL, returns the output list. Otherwise, writes results
to the specified path.
}
\description{
This function executes the transition risk stress test. It reads data from
the input path, runs the transition risk model using the provided parameters,
and writes results to the output path if provided.
}
