% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{st_read_agnostic}
\alias{st_read_agnostic}
\title{Read agnostic data from a directory}
\usage{
st_read_agnostic(dir)
}
\arguments{
\item{dir}{A string representing the directory where the CSV files are stored.}
}
\value{
A list containing four elements: \code{scenarios_data}, \code{financial_data}, \code{assets_data}, and \code{carbon_data}.
}
\description{
This function reads scenario, financial, assets, and carbon price data
from CSV files located in a specified directory.
}
