% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{run_trisk}
\alias{run_trisk}
\title{Run stress testing for provided asset type.}
\usage{
run_trisk(
  input_path,
  output_path = NULL,
  return_results = FALSE,
  baseline_scenario,
  shock_scenario,
  scenario_geography,
  start_year = 2022,
  carbon_price_model = "no_carbon_tax",
  ...
)
}
\arguments{
\item{input_path}{String holding path to project agnostic data.}

\item{output_path}{String holding path to which output files are written.
NOTE: Results and logs per run are saved to a subdirectory of output_path
that will be generated automatically. The name of the subdirectory is the
timestamp of the run of the analysis.}

\item{return_results}{Boolean, indicating if results shall be exported.}

\item{baseline_scenario}{Holds the name of the baseline scenario to be used
in the stress test, for accepted value range check \code{stress_test_arguments}.}

\item{shock_scenario}{Holds the name of the shock scenario to be used in the
stress test, for accepted value range check \code{stress_test_arguments}.}

\item{scenario_geography}{Character vector, indicating which geographical
region(s) (concerning asset location) results shall be calculated for. For
accepted values compare \code{stress_test_arguments}.}

\item{start_year}{Numeric, first year in the analysis used as the starting
point from which production forecasts are compared against scenario targets.
Must be available in the production data and indicates the first year of
the scenario data.}

\item{carbon_price_model}{Character vector, indicating which NGFS model is used in regards to
carbon prices. Default is no carbon tax.}
}
\description{
This function runs the transition risk stress test. It can be desirable to
understand sensitivities of the scenarios, in which case the user may pass a
vector of values to one (and only one) of the detail arguments. This will
result in running the analysis multiple times in a row with the argument
varied.
NOTE: if \code{return_results} is TRUE results will not be written to \verb{output path} but instead are returned.
}
