% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{run_trisk}
\alias{run_trisk}
\title{Run stress testing for provided asset type.}
\usage{
run_trisk(input_path, output_path = NULL, show_params_cols = TRUE, ...)
}
\arguments{
\item{input_path}{String containing the path to project-agnostic data.}

\item{output_path}{String containing the path to which output files are written.
NOTE: Results and logs per run are saved to a subdirectory of output_path
that will be generated automatically. The subdirectory name is the
timestamp of the analysis run.}

\item{show_params_cols}{Logical, indicating whether to display parameter columns in the output.}
}
\value{
If output_path is NULL, returns the output_list. Otherwise, writes results to the specified path.
}
\description{
This function executes the transition risk stress test. It can be beneficial to
understand scenario sensitivities, in which case the user may pass a
vector of values to one (and only one) of the detailed arguments. This will
result in running the analysis multiple times consecutively with the varied argument.
NOTE: If \code{return_results} is TRUE, results will not be written to the \code{output_path}
but will instead be returned.
}
