% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{run_trisk}
\alias{run_trisk}
\title{Run stress testing for provided asset type.}
\usage{
run_trisk(
  input_path,
  output_path = NULL,
  save_and_check = TRUE,
  show_params_cols = TRUE,
  ...
)
}
\arguments{
\item{input_path}{String holding path to project agnostic data.}

\item{output_path}{String holding path to which output files are written.
NOTE: Results and logs per run are saved to a subdirectory of output_path
that will be generated automatically. The name of the subdirectory is the
timestamp of the run of the analysis.}

\item{save_and_check}{Boolean, indicating if results shall be exported.}
}
\description{
This function runs the transition risk stress test. It can be desirable to
understand sensitivities of the scenarios, in which case the user may pass a
vector of values to one (and only one) of the detail arguments. This will
result in running the analysis multiple times in a row with the argument
varied.
NOTE: if \code{return_results} is TRUE results will not be written to \verb{output path} but instead are returned.
}
