% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk.R
\name{run_trisk_model}
\alias{run_trisk_model}
\title{Run transition risk stress test model}
\usage{
run_trisk_model(
  assets_data,
  scenarios_data,
  financial_data,
  carbon_data,
  baseline_scenario = "",
  target_scenario = "",
  scenario_geography = "Global",
  carbon_price_model = "no_carbon_tax",
  lgd = 0.45,
  risk_free_rate = 0.02,
  discount_rate = 0.07,
  growth_rate = 0.03,
  div_netprofit_prop_coef = 1,
  shock_year = 2030,
  market_passthrough = 0
)
}
\arguments{
\item{assets_data}{Data frame containing asset information.}

\item{scenarios_data}{Data frame containing scenario information.}

\item{financial_data}{Data frame containing financial information.}

\item{carbon_data}{Data frame containing carbon price information.}

\item{baseline_scenario}{String specifying the name of the baseline scenario.}

\item{target_scenario}{String specifying the name of the shock scenario.}

\item{scenario_geography}{Character vector indicating which geographical region(s) to calculate results for.}

\item{carbon_price_model}{Character vector specifying which NGFS model to use for carbon prices. Default is "no_carbon_tax".}

\item{lgd}{Numeric value for Loss Given Default. Default is 0.45.}

\item{risk_free_rate}{Numeric value for the risk-free interest rate. Default is 0.02.}

\item{discount_rate}{Numeric value for the discount rate of dividends per year in the DCF. Default is 0.07.}

\item{growth_rate}{Numeric value for the terminal growth rate of profits beyond the final year in the DCF. Default is 0.03.}

\item{div_netprofit_prop_coef}{Numeric coefficient determining how strongly future dividends propagate to company value. Default is 1.}

\item{shock_year}{Numeric value specifying the year when the shock is applied. Default is 2030.}

\item{market_passthrough}{Numeric value representing the firm's ability to pass carbon tax onto the consumer. Default is 0.}
}
\value{
A list containing:
\item{company_pd_changes_overall}{Data frame of overall probability of default changes}
\item{company_trajectories}{Data frame of company annual profits}
\item{company_technology_npv}{Data frame of company technology net present values}
}
\description{
This function executes the core transition risk stress test model calculations.
}
